pyproject.toml
[project]
name = "camera-recording-index"
version = "0.1.0"
description = ""
authors = [
    { name = "Anders Birkenes", email = "datagutten@datagutten.net" }
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow (>=11.2.1,<12.0.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)",
    "pytz (>=2025.2,<2026.0)",
    "django (>=5.2.1,<6.0.0)",
    "django-bootstrap5 (>=25.1,<26.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.reolink.dependencies]
reolinkapi = "^0.4.1"


[tool.poetry.group.gunicorn.dependencies]
gunicorn = "^23.0.0"


[tool.poetry.group.dotenv.dependencies]
dotenv = "^0.9.9"


[tool.poetry.group.dev.dependencies]
django-stubs = "^5.2.0"

